requirements.txt
watchdog
faster-whisper
openai
python-dotenv
